feat: add description to VIP-288
