+++ simulations/vip-288/abi/vTokenCorePool.json
